% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleSpreadsheetAccess.R
\name{access.google.spreadsheet}
\alias{access.google.spreadsheet}
\title{Retrieve Google Spreadsheet Workbook.}
\usage{
access.google.spreadsheet(docname)
}
\arguments{
\item{docname}{name of the spreadsheet}
}
\description{
Retrieve Google Spreadsheet Workbook.
}

