% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSSpendNLPAnalytics.R
\name{categories.tohash}
\alias{categories.tohash}
\title{Put categories (stemmed) into a global hash table.}
\usage{
categories.tohash(categories)
}
\arguments{
\item{categories}{list of categories}
}
\value{
hash map
}
\description{
Put categories (stemmed) into a global hash table.
}
