% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSSpendNLPAnalytics.R
\name{choose.bestword}
\alias{choose.bestword}
\title{Choose the best word (highest score).}
\usage{
choose.bestword(words)
}
\arguments{
\item{words}{list of words}
}
\value{
the word of the highest score
}
\description{
Choose the best word (highest score).
}

