% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSSpendAnalytics.R
\name{generate.summary}
\alias{generate.summary}
\title{Calculate the summary and output a dataframe}
\usage{
generate.summary(wb, pausing.time = 6)
}
\arguments{
\item{wb}{Google spreadsheet workbook}

\item{pausing.time}{pausing time in retrieving data, in seconds (default: 6)}
}
\value{
summary worksheet
}
\description{
Calculate the summary and output a dataframe
}

