% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpendingDataRetrieval.R
\name{get.all.spend.data}
\alias{get.all.spend.data}
\title{Retrieve all spend data}
\usage{
get.all.spend.data(wb, pausing.time = 6)
}
\arguments{
\item{wb}{Google Spreadsheet workbook}

\item{pausing.time}{pausing time in retrieving data, in seconds (default: 6)}
}
\value{
all spending data frame
}
\description{
Retrieve all spend data
}
