% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatLangProc.R
\name{get.postag.annotator}
\alias{get.postag.annotator}
\title{Return the singleton of part-of-speech (POS) tagging annotator}
\usage{
get.postag.annotator()
}
\value{
part-of-speech tagging annotator
}
\description{
Return the singleton of part-of-speech (POS) tagging annotator
}
