% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpendingDataRetrieval.R
\name{get.spend.data.months}
\alias{get.spend.data.months}
\title{Retrieve spending data of a few given months from the given workbook. Dataframe will be concatenated.}
\usage{
get.spend.data.months(wb, months)
}
\arguments{
\item{wb}{Google Spreadsheet workbook}

\item{months}{a vector of months}
}
\value{
concatenated data frame of the spending of the given months
}
\description{
Retrieve spending data of a few given months from the given workbook. Dataframe will be concatenated.
}
