% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpendingDataRetrieval.R
\name{get.spend.data}
\alias{get.spend.data}
\title{Retrieve spending data of a given month from the given workbook.}
\usage{
get.spend.data(wb, month)
}
\arguments{
\item{wb}{Google Spreadsheet workbook}

\item{month}{Month ('January', 'Feburary' ...)}
}
\value{
data frame of the spending of the given month
}
\description{
Retrieve spending data of a given month from the given workbook.
}

