% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatLangProc.R
\name{get.token.annotator}
\alias{get.token.annotator}
\title{Return the singleton of maxent word token annotator}
\usage{
get.token.annotator()
}
\value{
maxent word token annotator
}
\description{
Return the singleton of maxent word token annotator
}
