% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSSpendNLPAnalytics.R
\name{normalize.category}
\alias{normalize.category}
\title{Normalize the category.}
\usage{
normalize.category(category, all.hash, xwalk.map)
}
\arguments{
\item{category}{category to be normalized}

\item{all.hash}{hash table for all categories}

\item{xwalk.map}{crosswalk hash map}
}
\value{
normalized category
}
\description{
Normalize the category.
}

