% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSSpendAnalytics.R
\name{online.update.gs.summary}
\alias{online.update.gs.summary}
\title{Update summary}
\usage{
online.update.gs.summary(wb, summary.df)
}
\arguments{
\item{summary.df}{summary data frame}
}
\description{
Update summary
}
