% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatLangProc.R
\name{stem.all.tokens}
\alias{stem.all.tokens}
\title{Stem all tokens using Porter stemmer.}
\usage{
stem.all.tokens(string)
}
\arguments{
\item{string}{string of sentence to stem}
}
\value{
arrays of stemmed tokens
}
\description{
Stem all tokens using Porter stemmer.
}
