% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatLangProc.R
\name{tag.POS}
\alias{tag.POS}
\title{Tag the part-of-speech (POS) of the tokens in the given string sentence.}
\usage{
tag.POS(string)
}
\arguments{
\item{string}{sentence to tag}
}
\value{
arrays of part-of-speech (tags used in Penn Treebank Project)
}
\description{
Tag the part-of-speech (POS) of the tokens in the given string sentence.
}

